Mark `concat_str` as pure
